Making copy_images more robust

+++ docs/copy_images.sh
@@ -1,1 +1,6 @@
-cp design/*.png dox/html
+if [ -e dox/html ]; then
+    cp design/*.png dox/html
+fi
+if [ -e docs/html ]; then
+    cp design/*.png docs/html
+fi
